Move preprocessing in seperate package
