--- compoundsplitting/dictionary_pre_processing.go
@@ -1,1 +1,0 @@
-package compoundsplitting
