--- compoundsplitting/dictionary_pre_processing_test.go
@@ -1,1 +1,0 @@
-package compoundsplitting
